moved loggingService and logger out of core and into log. updated all references to loggingService and logger and set loggers to private. wall and player classes had an issue with entity having a private logger, so I deleted them as they were not being used yet anyways.
