+++ src/app/control/component/left-control/left-control.component.ts
@@ -2,8 +2,8 @@ import {
   Component, EventEmitter, Output
 } from '@angular/core';
 
-import { LoggingService } from '../../../core/service/logging.service';
-import { Logger } from '../../../core/model/logger';
+import { LoggingService } from '../../../log/service/logging.service';
+import { Logger } from '../../../log/model/logger';
 
 @Component({
   selector: 'left-control',
@@ -11,7 +11,7 @@ import { Logger } from '../../../core/model/logger';
   styleUrls: ['./left-control.component.css']
 })
 export class LeftControlComponent {
-  logger: Logger = LoggingService.getLogger('LeftControlComponent');
+  private logger: Logger = LoggingService.getLogger('LeftControlComponent');
 
   @Output() onControlPressed: EventEmitter<string>;
   width: number;
