+++ src/app/core/core.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LoggingService } from './service/logging.service';
 import { ApiService } from './service/api.service';
 
 @NgModule({
@@ -8,6 +7,6 @@ import { ApiService } from './service/api.service';
     CommonModule
   ],
   declarations: [],
-  providers: [ LoggingService, ApiService ]
+  providers: [ ApiService ]
 })
 export class CoreModule {}
