+++ src/app/log/log.module.ts
@@ -1,11 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+
 import { AppConsoleComponent } from './component/app-console/app-console.component';
+
+import { LoggingService } from './service/logging.service';
 
 @NgModule({
   imports: [
     CommonModule
   ],
-  declarations: [ AppConsoleComponent ]
+  declarations: [ AppConsoleComponent ],
+  providers: [ LoggingService]
 })
 export class LogModule { }
