+++ src/app/log/service/logging.service.ts
