+++ src/app/render/component/canvas/canvas.component.ts
@@ -2,8 +2,8 @@ import {
   Component, ViewChild, ElementRef, AfterViewInit
 } from '@angular/core';
 
-import { LoggingService } from '../../../core/service/logging.service';
-import { Logger } from '../../../core/model/logger';
+import { LoggingService } from '../../../log/service/logging.service';
+import { Logger } from '../../../log/model/logger';
 
 import { RenderService } from '../../service/render.service';
 import { Camera } from '../../model/camera';
@@ -14,7 +14,7 @@ import { Camera } from '../../model/camera';
   styleUrls: ['./canvas.component.css']
 })
 export class CanvasComponent implements AfterViewInit {
-  logger: Logger = LoggingService.getLogger('CanvasComponent');
+  private logger: Logger = LoggingService.getLogger('CanvasComponent');
 
   @ViewChild('canvas') canvas: ElementRef;
