+++ src/app/render/model/entity.ts
@@ -1,8 +1,8 @@
-import { LoggingService } from '../../core/service/logging.service';
-import { Logger } from '../../core/model/logger';
+import { LoggingService } from '../../log/service/logging.service';
+import { Logger } from '../../log/model/logger';
 
 export class Entity {
-  logger: Logger = LoggingService.getLogger('Entity');
+  private logger: Logger = LoggingService.getLogger('Entity');
 
   entityId: string;
