+++ src/app/render/model/plane.ts
@@ -1,10 +1,10 @@
-import { LoggingService } from '../../core/service/logging.service';
-import { Logger } from '../../core/model/logger';
+import { LoggingService } from '../../log/service/logging.service';
+import { Logger } from '../../log/model/logger';
 
 import { Vector3 } from './vector3';
 
 export class Plane {
-  logger: Logger = LoggingService.getLogger('Plane');
+  private logger: Logger = LoggingService.getLogger('Plane');
 
   normal: Vector3; // the normal vector of the plane
   distance: number; // the distance of the normal from the origin
